The result downloads, subtitle routes and clip cutting live in app/api/results.py and app/api/clips.py; fakes re-pointed, proven by canaries
